refactor: separate board database operations from lifecycle ownership
